Move org.slf4j/slf4j-log4j12 into :dev profile (fixes #1)
This dependency is superfluous as it is the binding between slf4j logging API and Log4j 1.2 logger. Thus it forces projects using this dependency to either use Log4j 1.2 or exclude this dependency.

The dependency is also scoped as runtime so it's not accessible to the code but available at runtime.