Use :provided profile to really fix #1
The dependencies of :dev profile are added to the generated pom.xml project descriptor. The content of :provided profile is added too, but scoped as "provided". It's a bit hacky but does the work.

For more details about dependency scopes, see https://maven.apache.org/guides/introduction/introduction-to-dependency-mechanism.html#Dependency_Scope